# Add peephole optimisation for joins