--- a/src/parse/asp/parser.go
+++ b/src/parse/asp/parser.go
@@ -222,6 +222,22 @@ func (p *Parser) optimise(statements []*Statement) []*Statement {
 		}
 		ret = append(ret, stmt)
 	}
+	// Spot ' '.join([...]) to be optimised later
+	WalkAST(ret, func(expr *Expression) bool {
+		if expr.Val != nil && expr.Val.String != "" && expr.Val.Property != nil && expr.Val.Property.Name == "join" && len(expr.Val.Property.Action) == 1 {
+			if call := expr.Val.Property.Action[0].Call; call != nil && len(call.Arguments) == 1 {
+				if arg := call.Arguments[0]; arg.Name == "" && arg.Value.Val != nil && arg.Value.Val.List != nil {
+					expr.optimised = &optimisedExpression{Join: &optimisedJoin{
+						Base: expr.Val.String,
+						Args: arg.Value.Val.List,
+					}}
+					expr.Val = nil
+					return false
+				}
+			}
+		}
+		return true
+	})
 	return ret
 }