Specifically cases like `' '.join([x for x in y if z])`

This comes up surprisingly often (>300k examples in an internal repo parse). This attempts to optimise it by avoiding creating throwaway lists and going straight into a `strings.Builder`. It seems to be about 30-40% faster which is quite a bit more than I expected (although I need to do some investigation internally, overall the numbers seem worse than I recall, but that is unrelated to this really).

Doesn't optimise anything with e.g. inline ifs which don't come up as often. There are cases but I don't think it's worth trying to contort the code to understand those right now.

```
$ plz run //tools/performance:parse_perf_test -- --plz plz 
INFO: Run 1 of 5
INFO: Complete in 29.73s, using 9797212 KB
INFO: Run 2 of 5
INFO: Complete in 29.56s, using 12509832 KB
INFO: Run 3 of 5
INFO: Complete in 35.13s, using 14891672 KB
INFO: Run 4 of 5
INFO: Complete in 42.46s, using 12615532 KB
INFO: Run 5 of 5
INFO: Complete in 33.68s, using 12739648 KB
INFO: Complete, median time: 33.68s, median mem: 12615532.00 KB
$ plz run //tools/performance:parse_perf_test -- --plz plz-out/bin/src/please
INFO: Run 1 of 5
INFO: Complete in 24.29s, using 13575496 KB
INFO: Run 2 of 5
INFO: Complete in 21.04s, using 12260728 KB
INFO: Run 3 of 5
INFO: Complete in 21.31s, using 12553096 KB
INFO: Run 4 of 5
INFO: Complete in 21.87s, using 14087104 KB
INFO: Run 5 of 5
INFO: Complete in 23.22s, using 12478676 KB
INFO: Complete, median time: 21.87s, median mem: 12553096.00 KB
```